pyproject.toml
[project]
name = "nonebot-plugin-shiro-signin"
version = "0.1.10"
description = "A NoneBot2 plugin for sign-in and favorability system themed with Shiro."
authors = [
    { name = "luojisama", email = "2534316454@qq.com" },
]
dependencies = [
    "nonebot2>=2.2.0",
    "nonebot-adapter-onebot>=2.4.0",
    "nonebot-plugin-htmlrender>=0.2.0",
    "nonebot-plugin-localstore>=0.6.0",
    "httpx>=0.23.0",
    "pydantic",
]
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
keywords = ["nonebot", "nonebot2", "plugin", "signin", "shiro"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Communications :: Chat",
]

[project.urls]
Homepage = "https://github.com/luojisama/nonebot-plugin-shiro-signin"
Repository = "https://github.com/luojisama/nonebot-plugin-shiro-signin"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nonebot_plugin_shiro_signin"]

[tool.hatch.build.targets.sdist]
include = [
    "/nonebot_plugin_shiro_signin",
]

[tool.nonebot]
plugins = ["nonebot_plugin_shiro_signin"]
